+++ lib/shelly/user.rb
@@ -37,14 +37,6 @@ module Shelly
       shelly.token["token"]
     end
 
-    def send_invitation(cloud, email)
-      shelly.send_invitation(cloud, email)
-    end
-
-    def delete_collaboration(cloud, email)
-      shelly.delete_collaboration(cloud, email)
-    end
-
     def load_credentials
       return unless credentials_exists?
       @email, @password = File.read(credentials_path).split("\n")