+++ spec/shelly/client_spec.rb
@@ -249,7 +249,7 @@ describe Shelly::Client do
 
   describe "#send_invitation" do
     it "should send post with developer's email" do
-      FakeWeb.register_uri(:post, api_url("apps/staging-foo/collaborations"), :body => {}.to_json)
+      FakeWeb.register_uri(:post, api_url("organizations/staging-foo/memberships"), :body => {}.to_json)
       response = @client.send_invitation("staging-foo", "megan@example.com")
       response.should == {}
     end
@@ -257,8 +257,8 @@ describe Shelly::Client do
 
   describe "#delete_collaboration" do
     it "should send delete with developer's email in url" do
-      FakeWeb.register_uri(:delete, api_url("apps/staging-foo/collaborations/megan@example.com"), :body => {}.to_json)
-      @client.delete_collaboration("staging-foo", "megan@example.com")
+      FakeWeb.register_uri(:delete, api_url("organizations/staging-foo/memberships/megan@example.com"), :body => {}.to_json)
+      @client.delete_member("staging-foo", "megan@example.com")
     end
   end