+++ spec/shelly/organization_spec.rb
@@ -15,4 +15,64 @@ describe Shelly::Organization do
       @organization.apps
     end
   end
+
+  describe "#memberships" do
+    before do
+      response = [{"email" => "bb@example.com", "active" => true, "owner" => true},
+                  {"email" => "aa@example.com", "active" => true, "owner" => false},
+                  {"email" => "cc@example.com", "active" => false, "owner" => true},
+                  {"email" => "dd@example.com", "active" => false, "owner" => false}]
+      @client.stub(:members).with("foo-org").and_return(response)
+    end
+
+    it "should fetch organization's users" do
+      @client.should_receive(:members).with("foo-org")
+      @organization.memberships
+    end
+
+    it "should sort members by email" do
+      members = @organization.memberships
+      members.should == [{"email"=>"aa@example.com", "active"=>true, "owner"=>false},
+                         {"email"=>"bb@example.com", "active"=>true, "owner"=>true},
+                         {"email"=>"cc@example.com", "active"=>false, "owner"=>true},
+                         {"email"=>"dd@example.com", "active"=>false, "owner"=>false}]
+    end
+
+
+    context "owners" do
+      it "should return only owners without inactive members" do
+        owners = @organization.owners
+        owners.should == [{"email"=>"bb@example.com", "active"=>true, "owner"=>true}]
+      end
+    end
+
+    context "members" do
+      it "should return only members without inactive members" do
+        members = @organization.members
+        members.should == [{"email"=>"aa@example.com", "active"=>true, "owner"=>false}]
+      end
+    end
+
+    context "inactive_members" do
+      it "should return only inactive members" do
+        inactive = @organization.inactive_members
+        inactive.should == [{"email"=>"cc@example.com", "active"=>false, "owner"=>true},
+                           {"email"=>"dd@example.com", "active"=>false, "owner"=>false}]
+      end
+    end
+  end
+
+  describe "#send_invitation" do
+    it "should send invitation" do
+      @client.should_receive(:send_invitation).with("foo-org", "megan@example.com", true)
+      @organization.send_invitation("megan@example.com", true)
+    end
+  end
+
+  describe "#delete_member" do
+    it "should delete collaboration" do
+      @client.should_receive(:delete_member).with("foo-org", "megan@example.com")
+      @organization.delete_member("megan@example.com")
+    end
+  end
 end